move default terminal to shared crate
