+++ lazycard/src/lib.rs
@@ -2,7 +2,6 @@ pub mod app;
 pub mod database;
 pub mod pages;
 pub mod settings;
-pub mod terminal;
 
 pub const APP_NAME: &str = env!("CARGO_PKG_NAME");
 pub const APP_VERSION: &str = env!("CARGO_PKG_VERSION");
