+++ trollstov/src/lib.rs
@@ -5,7 +5,6 @@ pub mod jukebox;
 pub mod logo;
 pub mod pages;
 pub mod settings;
-pub mod terminal;
 
 const APP_NAME: &str = env!("CARGO_PKG_NAME");
 const APP_VERSION: &str = env!("CARGO_PKG_VERSION");
